Move application.cpp to new application folder
