+++ src/solver/signal-handling/linux.cpp
@@ -2,6 +2,7 @@
 #include <csignal>
 
 #include "common.h"
+#include "application/application.h"
 #include "public.h"
 
 namespace Antares::Solver
